Keyboard didn't always show up when starting a game on Android >= 7.0
Globally we have the soft keyboard set to stateAlwaysVisible. When
setContentView is called in startTerp this appears to cause a race
condition that can make the keyboard disappear depending on timing.
The keyboard being shown is dependent on the presence of focusable
views. In a fresh GlkLayout there are none. The terp thread will
schedule tasks on the ui thread to create the windows it needs and
it appears that the keyboard being shown or not is dependent on in
what order certain events end up in the ui thread queue. I haven't
been able to figure out yet why this is.

To avoid the issue I've removed the repeated setContentView calls and
put the glkLayout in the main ll layout. To switch between the
welcome screen and glk view the appropriate views are hidden and
shown.